test(drive): sibling file for the time-range e2e module; pin cursor ids and the unique layout
- Move the 1,550-line time_range_index_e2e_tests module out of
  add_document_for_contract/mod.rs into a sibling file, leaving the
  64-line versioned dispatcher readable.
- The in-bucket cursor regression now asserts the exact returned id
  sequence (deserialized), not just result counts.
- New coverage for the unique arm of the in-bucket cursor rule: on a
  unique single-property grid an included cursor retains the bucket's
  sole document and an excluded cursor empties the page.
